Refactor all module files
